requirements.txt
anthropic>=0.43.1
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
schedule>=1.2.1
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.9
jinja2>=3.1.4
aiofiles>=23.2.1
